+++ examples/cleanup/playwright.config.ts
@@ -3,7 +3,13 @@ import { globalCache } from '@global-cache/playwright';
 
 const config = defineConfig({
   testDir: './test',
-  globalTeardown: require.resolve('./test/cleanup'), // <-- custom teardown script for cleanup
 });
 
-export default globalCache.wrap(config);
+export default globalCache.wrap(config, {
+  cleanup: async () => {
+    const userId = await globalCache.getStale('db-user-id');
+    if (userId) {
+      console.log(`\nRemoving user from db: ${userId}\n`);
+    }
+  },
+});