+++ examples/cleanup/test/cleanup.ts
@@ -1,8 +1,0 @@
-import { globalCache } from '@global-cache/playwright';
-
-export default async function cleanup() {
-  const userId = await globalCache.getStale('db-user-id');
-  if (userId) {
-    console.log(`\nRemoving user from db: ${userId}\n`);
-  }
-}