+++ packages/playwright/test/playwright.config.ts
@@ -5,4 +5,11 @@ const config = defineConfig({
   testDir: '.',
 });
 
-export default globalCache.wrap(config);
+export default globalCache.wrap(config, {
+  cleanup: async () => {
+    const userId = await globalCache.getStale('user-id');
+    if (userId !== 'foo') {
+      throw new Error(`Expected stale userId to be "foo", got: ${JSON.stringify(userId)}`);
+    }
+  },
+});